fix import in unit test
