+++ tests/maintenance/conftest.py
@@ -16,6 +16,29 @@ from unittest import mock
 
 import pytest
 
+from antarest.core.config import Config
+
+
+def load_config_mock() -> None:
+    pass
+
+
+def get_config_mock() -> Config:
+    return Config()
+
+
+# Explanation:
+#   The Celery app setup requires to initialize it at the module level.
+#   Here in tests, we override the initialization functions used in prod, otherwise they will raise because
+#   they require the env var ANTARES_CONF to be filled with a path to a working configuration.
+#
+#   We perform that import here in conftest, so that all tests in the directory do not need to do it again.
+with (
+    mock.patch("antarest.maintenance.config.load_config", load_config_mock),
+    mock.patch("antarest.maintenance.config.get_config", get_config_mock),
+):
+    import antarest.maintenance.app  # noqa
+
 
 @pytest.fixture
 def with_no_maintenance_ctx():
